added babel & eslint
