+++ BackEnd/package.json
@@ -4,17 +4,34 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon server.js",
+    "start": "nodemon ./src/server.js --exec babel-node -e js",
+    "transpile": "./node_modules/.bin/babel src -d dist",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "Caleb Reiter",
+  "author": "Nick Lanman",
+  "contributers": [
+    "Caleb Reiter"
+  ],
   "license": "ISC",
   "dependencies": {
+    "core-js": "^3.5.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "multer": "^1.4.2"
+    "multer": "^1.4.2",
+    "regenerator-runtime": "^0.13.3"
   },
   "devDependencies": {
+    "@babel/cli": "^7.7.5",
+    "@babel/core": "^7.7.5",
+    "@babel/node": "^7.7.4",
+    "@babel/preset-env": "^7.7.6",
+    "babel-eslint": "^10.0.3",
+    "eslint": "^6.7.2",
     "nodemon": "^2.0.2"
+  },
+  "babel": {
+    "presets": [
+      "@babel/env"
+    ]
   }
 }
