+++ BackEnd/src/helpers/multer.js
@@ -1,15 +1,16 @@
-const multer = require('multer');
+import multer from 'multer';
 
 const resumeStorage = (() => {
   const destination = process.env.DIR_RESUME || './MockData/Resumes';
   return multer.diskStorage({
     destination,
     filename: (req, file, cb) => {
-        const filename = `${req.params.id}.pdf`;
-        cb(null, filename);
-      }
-    })
+      const filename = `${req.params.id}.pdf`;
+      cb(null, filename);
+    }
+  });
 })();
+
 const resumeUpload = multer({
   storage: resumeStorage,
   fileFilter: (req, file, cb) => {
